+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# v1.1.1
+##  11/17/2016
+
+1. [](#improved)
+    * Removed dependency of PECL YAML function yaml_emit() in favor of using vendor/spyc.php class.
+
+# v1.1.0
+##  11/16/2016
+
+1. [](#new)
+    * Settings in the pagefrontmatter block in the form page frontmatter now are merged with values from form fields. Form field values ovverride the pagefrontmatter settings.
+
+# v1.0.0
+##  11/15/2016
+
+1. [](#new)
+    * Added an extra form field: 'author'
+    * Added copying an (optional) frontmatter block from the form page frontmatter to the newly added page's frontmatter
+
+# v0.2.0
+##  11/13/2016
+
+1. [](#new)
+    * Plugin name changed to Add Page By Form (add-page-by-form)
+    * Added timestamp as date in page header (date format is taken from plugin config)
+    * Pages with identical titles are saved by adding a incremental number to the page slug (e.g. 'my-page\_2', 'my-page\_2', etc.)
+    * Added error handling
+
 # v0.1.0
 ##  11/08/2016